console/requirements.txt
openai
azure-cosmos